feat(ai): Abstrahiere Kontextkonstruktion in IChatContextBuilder
