+++ src/ghGPT.Core/Ai/IChatContextBuilder.cs
@@ -1,0 +1,6 @@
+namespace ghGPT.Core.Ai;
+
+public interface IChatContextBuilder
+{
+    Task<IEnumerable<ChatMessage>> BuildAsync(ChatRequest request);
+}
